+++ components/display/display.c
@@ -28,6 +28,9 @@ void display_init(void)
 #if CONFIG_DISPLAY_ILI9341
     extern void ili9341_init(void);
     ili9341_init();
+#elif CONFIG_DISPLAY_ST7789V
+    extern void st7789v_init(void);
+    st7789v_init();
 #elif CONFIG_DISPLAY_CO5300
     extern void co5300_init(void);
     co5300_init();